`tools/importer-rest-api-specs`: moving the Terraform Schema Generation package into the Terraform package
