+++ tools/importer-rest-api-specs/components/terraform/schema/builder.go
@@ -6,9 +6,9 @@ package schema
 import (
 	"fmt"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/terraform/helpers"
+	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/terraform/schema/processors"
 
 	"github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/schema/processors"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/models"
 	"github.com/hashicorp/pandora/tools/sdk/resourcemanager"
 )
