+++ tools/importer-rest-api-specs/components/terraform/schema/fields_resource_id.go
