+++ tools/importer-rest-api-specs/components/terraform/schema/helpers.go
@@ -6,11 +6,11 @@ package schema
 import (
 	"fmt"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/terraform/helpers"
+	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/terraform/schema/processors"
 	"regexp"
 	"strings"
 	"unicode"
 
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/schema/processors"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/models"
 	"github.com/hashicorp/pandora/tools/sdk/resourcemanager"
 )
