+++ tools/importer-rest-api-specs/components/terraform/schema/object_definition.go
