+++ tools/importer-rest-api-specs/components/terraform/schema/processors/field_name_plural_to_singular.go
