+++ tools/importer-rest-api-specs/components/terraform/schema/processors/field_name_remove_is_prefix_test.go
