+++ tools/importer-rest-api-specs/components/terraform/schema/processors/field_name_remove_resource_prefix_test.go
