+++ tools/importer-rest-api-specs/components/terraform/schema/processors/field_name_rename_boolean_test.go
