+++ tools/importer-rest-api-specs/components/terraform/schema/processors/helpers.go
