+++ tools/importer-rest-api-specs/components/terraform/schema/processors/model_flatten_properties_into_parent.go
