+++ tools/importer-rest-api-specs/components/terraform/schema/processors/model_flatten_reference_id.go
