+++ tools/importer-rest-api-specs/components/terraform/schema/processors/model_flatten_sku_name_test.go
