+++ tools/importer-rest-api-specs/components/terraform/schema/processors/model_remove_status_detail.go
