+++ tools/importer-rest-api-specs/components/terraform/schema/processors/model_rename_zones_test.go
